// packages/kratos-cdk/models/Message.ts
export class Message {
    'body': string;
    /**
    * CreatedAt is a helper struct field for gobuffalo.pop.
    */
    'created_at': Date;
    /**
    * Dispatches store information about the attempts of delivering a message May contain an error if any happened, or just the `success` state.
    */
    'dispatches'?: Array<MessageDispatch>;
    'id': string;
    'recipient': string;
    'send_count': number;
    'status': CourierMessageStatus;
    'subject': string;
    /**
    *  recovery_invalid TypeRecoveryInvalid recovery_valid TypeRecoveryValid recovery_code_invalid TypeRecoveryCodeInvalid recovery_code_valid TypeRecoveryCodeValid verification_invalid TypeVerificationInvalid verification_valid TypeVerificationValid verification_code_invalid TypeVerificationCodeInvalid verification_code_valid TypeVerificationCodeValid otp TypeOTP stub TypeTestStub
    */
    'template_type': MessageTemplateTypeEnum;
    'type': CourierMessageType;
    /**
    * UpdatedAt is a helper struct field for gobuffalo.pop.
    */
    'updated_at': Date;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "body",
            "baseName": "body",
            "type": "string",
            "format": ""
        },
        {
            "name": "created_at",
            "baseName": "created_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "dispatches",
            "baseName": "dispatches",
            "type": "Array<MessageDispatch>",
            "format": ""
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string",
            "format": "uuid"
        },
        {
            "name": "recipient",
            "baseName": "recipient",
            "type": "string",
            "format": ""
        },
        {
            "name": "send_count",
            "baseName": "send_count",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "CourierMessageStatus",
            "format": ""
        },
        {
            "name": "subject",
            "baseName": "subject",
            "type": "string",
            "format": ""
        },
        {
            "name": "template_type",
            "baseName": "template_type",
            "type": "MessageTemplateTypeEnum",
            "format": ""
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "CourierMessageType",
            "format": ""
        },
        {
            "name": "updated_at",
            "baseName": "updated_at",
            "type": "Date",
            "format": "date-time"
        }    ];

    static getAttributeTypeMap() {
        return Message.attributeTypeMap;
    }

    public constructor() {
    }
}


export type MessageTemplateTypeEnum = "recovery_invalid" | "recovery_valid" | "recovery_code_invalid" | "recovery_code_valid" | "verification_invalid" | "verification_valid" | "verification_code_invalid" | "verification_code_valid" | "otp" | "stub" ;